[core] use uint32_t in http_header.[ch]
headers should never approach anywhere remotely close to > 4 GB

+++ src/http_header.h
@@ -41,27 +41,27 @@ enum http_header_e {
 };
 
 __attribute_pure__
-enum http_header_e http_header_hkey_get(const char *s, size_t slen);
+enum http_header_e http_header_hkey_get(const char *s, uint32_t slen);
 
 __attribute_pure__
-int http_header_str_contains_token (const char *s, size_t slen, const char *m, size_t mlen);
+int http_header_str_contains_token (const char *s, uint32_t slen, const char *m, uint32_t mlen);
 
 __attribute_pure__
-buffer * http_header_response_get(const request_st *r, enum http_header_e id, const char *k, size_t klen);
-void http_header_response_unset(request_st *r, enum http_header_e id, const char *k, size_t klen);
-void http_header_response_set(request_st *r, enum http_header_e id, const char *k, size_t klen, const char *v, size_t vlen);
-void http_header_response_append(request_st *r, enum http_header_e id, const char *k, size_t klen, const char *v, size_t vlen);
-void http_header_response_insert(request_st *r, enum http_header_e id, const char *k, size_t klen, const char *v, size_t vlen);
+buffer * http_header_response_get(const request_st *r, enum http_header_e id, const char *k, uint32_t klen);
+void http_header_response_unset(request_st *r, enum http_header_e id, const char *k, uint32_t klen);
+void http_header_response_set(request_st *r, enum http_header_e id, const char *k, uint32_t klen, const char *v, uint32_t vlen);
+void http_header_response_append(request_st *r, enum http_header_e id, const char *k, uint32_t klen, const char *v, uint32_t vlen);
+void http_header_response_insert(request_st *r, enum http_header_e id, const char *k, uint32_t klen, const char *v, uint32_t vlen);
 
 __attribute_pure__
-buffer * http_header_request_get(const request_st *r, enum http_header_e id, const char *k, size_t klen);
-void http_header_request_unset(request_st *r, enum http_header_e id, const char *k, size_t klen);
-void http_header_request_set(request_st *r, enum http_header_e id, const char *k, size_t klen, const char *v, size_t vlen);
-void http_header_request_append(request_st *r, enum http_header_e id, const char *k, size_t klen, const char *v, size_t vlen);
+buffer * http_header_request_get(const request_st *r, enum http_header_e id, const char *k, uint32_t klen);
+void http_header_request_unset(request_st *r, enum http_header_e id, const char *k, uint32_t klen);
+void http_header_request_set(request_st *r, enum http_header_e id, const char *k, uint32_t klen, const char *v, uint32_t vlen);
+void http_header_request_append(request_st *r, enum http_header_e id, const char *k, uint32_t klen, const char *v, uint32_t vlen);
 
 __attribute_pure__
-buffer * http_header_env_get(const request_st *r, const char *k, size_t klen);
-void http_header_env_set(request_st *r, const char *k, size_t klen, const char *v, size_t vlen);
-void http_header_env_append(request_st *r, const char *k, size_t klen, const char *v, size_t vlen);
+buffer * http_header_env_get(const request_st *r, const char *k, uint32_t klen);
+void http_header_env_set(request_st *r, const char *k, uint32_t klen, const char *v, uint32_t vlen);
+void http_header_env_append(request_st *r, const char *k, uint32_t klen, const char *v, uint32_t vlen);
 
 #endif
